finish all modules and readme
