+++ mmrotate_sam/data_builder.py
@@ -5,6 +5,8 @@ from typing import Dict, Optional
 
 from torch.utils.data import DataLoader
 
+from mmcv.transforms import BaseTransform
+
 from mmengine.evaluator import Evaluator
 from mmengine.dataset import worker_init_fn
 from mmengine.dist import get_rank
@@ -13,9 +15,7 @@ from mmengine.registry import DATA_SAMPLERS, FUNCTIONS, EVALUATOR
 from mmengine.utils import digit_version
 from mmengine.utils.dl_utils import TORCH_VERSION
 
-from mmrotate.registry import DATASETS
-
-import transforms
+from mmrotate.registry import DATASETS, TRANSFORMS
 
 
 def build_data_loader(data_name=None):
@@ -31,6 +31,27 @@ def build_evaluator(merge_patches=True, format_only=False):
     naive_evaluator.update(dict(
         merge_patches=merge_patches, format_only=format_only))
     return MMEngine_build_evaluator(evaluator=naive_evaluator)
+
+
+@TRANSFORMS.register_module()
+class AddConvertedGTBox(BaseTransform):
+    """Convert boxes in results to a certain box type."""
+
+    def __init__(self, box_type_mapping: dict) -> None:
+        self.box_type_mapping = box_type_mapping
+
+    def transform(self, results: dict) -> dict:
+        """The transform function."""
+        for key, dst_box_type in self.box_type_mapping.items():
+            assert key != 'gt_bboxes'
+            gt_bboxes = results['gt_bboxes']
+            results[key] = gt_bboxes.convert_to(dst_box_type)
+        return results
+
+    def __repr__(self):
+        repr_str = self.__class__.__name__
+        repr_str += f'(box_type_mapping={self.box_type_mapping})'
+        return repr_str
 
 
 # dataset settings
